requirements.txt
ultralytics>=8.2.0
opencv-python>=4.9.0
numpy>=1.26.0
PyYAML>=6.0.1
deep-sort-realtime>=1.3.2
scipy>=1.11.0
